+++ tools/CreateLnk.cs
@@ -1,0 +1,44 @@
+using System;
+using System.Runtime.InteropServices;
+using System.Runtime.InteropServices.ComTypes;
+
+[ComImport, Guid("00021401-0000-0000-C000-000000000046")]
+class ShellLink { }
+
+[ComImport, InterfaceType(ComInterfaceType.InterfaceIsIUnknown),
+ Guid("000214F9-0000-0000-C000-000000000046")]
+interface IShellLinkW
+{
+    void GetPath([Out, MarshalAs(UnmanagedType.LPWStr)] System.Text.StringBuilder f, int c, IntPtr d, int fl);
+    void GetIDList(out IntPtr p);
+    void SetIDList(IntPtr p);
+    void GetDescription([Out, MarshalAs(UnmanagedType.LPWStr)] System.Text.StringBuilder s, int c);
+    void SetDescription([MarshalAs(UnmanagedType.LPWStr)] string s);
+    void GetWorkingDirectory([Out, MarshalAs(UnmanagedType.LPWStr)] System.Text.StringBuilder s, int c);
+    void SetWorkingDirectory([MarshalAs(UnmanagedType.LPWStr)] string s);
+    void GetArguments([Out, MarshalAs(UnmanagedType.LPWStr)] System.Text.StringBuilder s, int c);
+    void SetArguments([MarshalAs(UnmanagedType.LPWStr)] string s);
+    void GetHotkey(out short h);
+    void SetHotkey(short h);
+    void GetShowCmd(out int s);
+    void SetShowCmd(int s);
+    void GetIconLocation([Out, MarshalAs(UnmanagedType.LPWStr)] System.Text.StringBuilder s, int c, out int i);
+    void SetIconLocation([MarshalAs(UnmanagedType.LPWStr)] string s, int i);
+    void SetRelativePath([MarshalAs(UnmanagedType.LPWStr)] string s, int r);
+    void Resolve(IntPtr h, int f);
+    void SetPath([MarshalAs(UnmanagedType.LPWStr)] string s);
+}
+
+class Program
+{
+    // Uso: CreateLnk.exe <ruta.lnk> <target.exe> [argumentos]
+    static void Main(string[] args)
+    {
+        var link = (IShellLinkW)new ShellLink();
+        link.SetPath(args[1]);
+        link.SetWorkingDirectory(System.IO.Path.GetDirectoryName(args[1]));
+        if (args.Length > 2) link.SetArguments(args[2]);
+        ((IPersistFile)link).Save(args[0], false);
+        Console.WriteLine("creado: " + args[0]);
+    }
+}